Delete and patch CRs
